Extract view-embedded business logic into services (StructureView, BinderDetailView, ProteinListView)
pdb_structure() and its full helper chain (~440 lines) had no request
dependency and were already reused by ProteinView.py and StructureExportView.py
via direct view-to-view imports -- move them into a new
tpweb/services/structure_summary.py, per tpw-django-patterns (views delegate
to services, not the reverse). StructureView.py now only keeps the actual
View class plus its request-handling helpers.

BinderDetailView's _compute_properties (RDKit physicochemical/drug-likeness
descriptors, no request dependency) moves into binder_summary.py -- the
service module that already owns the sibling _binder_table_properties
descriptor function and make_binder_svg -- as compute_binder_properties.

ProteinListView.get() was a single ~745-line method. Extract its four
largest self-contained chunks into private methods, matching the class's
own established convention (_build_table_rows, _build_filter_groups, etc.):
_build_filter_presets (saved-filter label building), _resolve_visible_columns
(column selection + score-dict/tcolumns build), _apply_structure_and_annotation_filters
(structure-source/EC/GO protein queryset filtering), and
_rank_and_sort_proteins (the per-protein score/ranking loop plus the
column/score/accession sort). get() itself now reads as a sequence of named
steps instead of one long unbroken block.
