+++ tpweb/tests/test_services.py
@@ -95,9 +95,9 @@ from tpweb.services.workspace import (
     get_workspace_session_value,
     set_workspace_session_value,
 )
+from tpweb.services.structure_summary import _annotated_site_label, _pocket_residue_overlap
 from tpweb.views.FormulaForm import FormulaForm
 from tpweb.views.IndexView import should_show_home_pipeline_panel
-from tpweb.views.StructureView import _annotated_site_label, _pocket_residue_overlap
 from tpweb.services.llm.base import Message, ToolCall, ToolDefinition, ToolResult
 from tpweb.services.llm.openai_provider import OpenAIProvider
